+++ CHANGES.md
@@ -14,6 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - rename mk-part command into mk
 - the dump command now includes "hol.ml"
 - do not print ocaml warnings while dumping proofs
+- do not export unused proofs (about 7%)
+- make hol2dk stat give the number of unused proofs
+
+### Fixed
+
+- README.md
+- output of hol2dk axm
+- output of hol2dk mk
 
 ## 0.0.0 (2023-11-08)